updated employee document home page to match employer: g
