+++ src/app/employer/documents/components/SimpleQueryPanel.tsx
@@ -39,6 +39,7 @@ interface SimpleQueryPanelProps {
   styleOptions: Record<string, string>;
   referencePages: number[];
   setPdfPageNumber: (p: number) => void;
+  userRole?: 'employer' | 'employee';
 }
 
 const styleIcons: Record<string, React.ReactNode> = {
@@ -65,8 +66,10 @@ export function SimpleQueryPanel({
   setAiStyle,
   styleOptions,
   referencePages: _referencePages,
-  setPdfPageNumber: _setPdfPageNumber
+  setPdfPageNumber: _setPdfPageNumber,
+  userRole = 'employer',
 }: SimpleQueryPanelProps) {
+  const showCompanyScope = userRole === 'employer';
   const [isFocused, setIsFocused] = useState(false);
 
   return (
@@ -88,41 +91,43 @@ export function SimpleQueryPanel({
 
       <ScrollArea className="flex-1">
         <div className="p-5 space-y-5">
-          {/* Search Scope Toggle */}
-          <div className="space-y-2">
-            <span className="text-[10px] font-bold text-muted-foreground uppercase tracking-widest px-1 flex items-center gap-1.5">
-              <span className="w-1 h-1 rounded-full bg-violet-500" />
-              Search In
-            </span>
-            <div className="flex gap-2">
-              <button
-                onClick={() => setSearchScope('document')}
-                className={cn(
-                  "flex-1 flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl text-xs font-semibold transition-all duration-200",
-                  searchScope === 'document'
-                    ? "bg-violet-600 text-white shadow-lg shadow-violet-500/25"
-                    : "bg-slate-100 dark:bg-slate-800 text-slate-600 dark:text-slate-400 hover:bg-slate-200 dark:hover:bg-slate-700"
-                )}
-              >
-                <FileText className="w-3.5 h-3.5" />
-                This Document
-              </button>
-              <button
-                onClick={() => companyId && setSearchScope('company')}
-                disabled={!companyId}
-                className={cn(
-                  "flex-1 flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl text-xs font-semibold transition-all duration-200",
-                  searchScope === 'company'
-                    ? "bg-violet-600 text-white shadow-lg shadow-violet-500/25"
-                    : "bg-slate-100 dark:bg-slate-800 text-slate-600 dark:text-slate-400 hover:bg-slate-200 dark:hover:bg-slate-700",
-                  !companyId && "opacity-50 cursor-not-allowed"
-                )}
-              >
-                <Building2 className="w-3.5 h-3.5" />
-                All Documents
-              </button>
+          {/* Search Scope Toggle - only show company scope for employers */}
+          {showCompanyScope && (
+            <div className="space-y-2">
+              <span className="text-[10px] font-bold text-muted-foreground uppercase tracking-widest px-1 flex items-center gap-1.5">
+                <span className="w-1 h-1 rounded-full bg-violet-500" />
+                Search In
+              </span>
+              <div className="flex gap-2">
+                <button
+                  onClick={() => setSearchScope('document')}
+                  className={cn(
+                    "flex-1 flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl text-xs font-semibold transition-all duration-200",
+                    searchScope === 'document'
+                      ? "bg-violet-600 text-white shadow-lg shadow-violet-500/25"
+                      : "bg-slate-100 dark:bg-slate-800 text-slate-600 dark:text-slate-400 hover:bg-slate-200 dark:hover:bg-slate-700"
+                  )}
+                >
+                  <FileText className="w-3.5 h-3.5" />
+                  This Document
+                </button>
+                <button
+                  onClick={() => companyId && setSearchScope('company')}
+                  disabled={!companyId}
+                  className={cn(
+                    "flex-1 flex items-center justify-center gap-2 px-4 py-2.5 rounded-xl text-xs font-semibold transition-all duration-200",
+                    searchScope === 'company'
+                      ? "bg-violet-600 text-white shadow-lg shadow-violet-500/25"
+                      : "bg-slate-100 dark:bg-slate-800 text-slate-600 dark:text-slate-400 hover:bg-slate-200 dark:hover:bg-slate-700",
+                    !companyId && "opacity-50 cursor-not-allowed"
+                  )}
+                >
+                  <Building2 className="w-3.5 h-3.5" />
+                  All Documents
+                </button>
+              </div>
             </div>
-          </div>
+          )}
 
           {/* Response Style Pills */}
           <div className="space-y-2">
